move status log entry and header to separate modules
